refactor: refactor reward
