+++ webgenie/rewards/__init__.py
@@ -1,6 +1,6 @@
 from .reward import Reward
 from .visual_reward import VisualReward
 from .quality_reward import QualityReward
-from .rtc_reward import RtcReward
+from .rtc_reward.rtc_reward import RtcReward
 from .bert_reward import BertReward
 from .lighthouse_reward import LighthouseReward
