+++ webgenie/rewards/rtc_reward/__init__.py
@@ -1,0 +1,1 @@
+from .rtc_reward import RtcReward
