+++ webgenie/rewards/visual_reward/__init__.py
@@ -1,0 +1,1 @@
+from .visual_reward import VisualReward
